# Faster rational creation