Optimization of the creation of `Rational`. Partially addressing #390.

Benchmarked using 

```python
In [1]: %%timeit a=random.randint(1, 1e6); b=random.randint(1,1e6)
   ...:          symengine.Rational(a, b)

In [2]: %%timeit a=random.randint(1, 1e12); b=random.randint(1,1e12)
   ...:          symengine.Rational(a, b)
```

For smaller numbers I got 2800 ns before this PR and 660ns after. For larger numbers I got 4230ns before this PR and 1710 after. (sympy need 240 ns for both examples)